+++ sdk/src/sandbox.ts
@@ -5,52 +5,12 @@ import * as pty from 'node-pty';
 import * as os from 'os';
 import { spawn, ChildProcess } from 'child_process';
 import * as fs from 'fs';
-import * as path from 'path';
 import { randomBytes } from "crypto";
 import { parse as semverParse } from 'semver';
 import { SandboxPolicy, SandboxingMethod, ContainerConfig, ContainmentType, ExperimentalBackends } from './types';
 import { findWxcExecutable, findLxcExecutable, getPlatformSupport } from './platform';
 import { FileLogger } from './logger';
-
-/**
- * Generate a timestamped log file path in the given directory.
- */
-function makeLogFilePath(dir: string): string {
-  const ts = new Date().toISOString().replace(/[:.]/g, '-').replace(/Z$/, '');
-  const suffix = randomBytes(3).toString('hex');
-  return path.join(dir, `mxc-diag-${ts}-${suffix}.log`);
-}
-
-/**
- * Sets up logging and resolves the executable path + args.
- * Shared by both PTY and non-PTY spawn paths.
- */
-function prepareSpawn(
-  config: ContainerConfig,
-  options: SandboxSpawnOptions,
-): { executablePath: string; args: string[]; logger?: FileLogger; logFile?: string; startTime: number } {
-  let logger: FileLogger | undefined;
-  let logFile: string | undefined;
-  const logDir = options.logDir ?? (options.debug ? path.join(os.tmpdir(), 'mxc-logs') : undefined);
-  if (logDir) {
-    logFile = makeLogFilePath(logDir);
-    logger = new FileLogger(logFile);
-    logger.log('info', 'mxc.log.created', { logFile });
-  }
-
-  const startTime = Date.now();
-  logger?.log('info', 'mxc.spawn.start', {
-    platform: os.platform(),
-    arch: os.arch(),
-    containment: config.containment,
-  });
-
-  const { executablePath, args } = resolveExecutableAndArgs(config, options);
-
-  logger?.log('info', 'mxc.binary.resolved', { resolved: !!executablePath });
-
-  return { executablePath, args, logger, logFile, startTime };
-}
+import { prepareSpawn } from './helper';
 
 const SUPPORTED_VERSION = '0.5.0-alpha';
 const MIN_VERSION = '0.4.0-alpha';
@@ -478,7 +438,7 @@ function spawnWithConfig(
   workingDirectory?: string,
   env?: { [key: string]: string | undefined },
 ): pty.IPty {
-  const { executablePath, args, logger, startTime } = prepareSpawn(config, options);
+  const { executablePath, args, logger, startTime } = prepareSpawn(config, options, resolveExecutableAndArgs);
 
   try {
     const ptyOpts: pty.IPtyForkOptions = {
@@ -639,7 +599,7 @@ export function spawnSandboxFromConfig(
   env?: { [key: string]: string | undefined }
 ): pty.IPty | ChildProcess {
   if (options.usePty === false) {
-    const { executablePath, args, logger, startTime } = prepareSpawn(config, options);
+    const { executablePath, args, logger, startTime } = prepareSpawn(config, options, resolveExecutableAndArgs);
     try {
       const child = spawn(executablePath, args, {
         cwd: workingDirectory || process.cwd(),
